QNX: Refactored bar descriptor editor
The Bar descriptor editor now uses the PanelWidget with PropertiesPanels
instead, to make it more similar to the rest of the Qt Creator UI.

Split the BarDescriptorEditorWidget class into multiple, smaller,
classes. Also moved the Entry Text & Images panel to the General tab,
as the Application tab was getting too crowded.

Change-Id: I691ccf3638aaad4b5f0ad0d040094f2144d2cec0
Reviewed-by: Tobias Hunger <tobias.hunger@digia.com>