QNX: Abort deployment if device is not in Development Mode
This adds a new deploy step (BlackBerryCheckDevModeStep), which
checks for Development Mode being set on the device. If not set,
the deployment stops early on in the process. Without this step,
the deployment would not fail until after uploading the package.

Moved blackberry-deploy into a constant, as it's used in
multiple places by now.

Change-Id: I813754108fb4be281e752b12ac56d4f0b302077d
Reviewed-by: Mehdi Fekari <mfekari@rim.com>
Reviewed-by: Rafael Roquetto <rafael.roquetto@kdab.com>
Reviewed-by: Nicolas Arnaud-Cormos <nicolas@kdab.com>