QNX: Refactored connection handling for BlackBerry devices
The blackberry-connect connection is now established at
Qt Creator startup, and taken down when Qt Creator shuts down.
The user also has the ability to disconnect manually in the
Device settings widget.

The connection status is shown in the Devices settings widget,
as well as in the project selector. And it is impossible run or
debug an application without having the device connected.

Change-Id: I64f2e48a647e6d850851d2f20d0ec7bbd5980ea5
Reviewed-by: Christian Kandeler <christian.kandeler@digia.com>
Reviewed-by: Nicolas Arnaud-Cormos <nicolas@kdab.com>
Reviewed-by: Mehdi Fekari <mfekari@rim.com>