EnvironmentWidget: Use HeaderViewStrecher
Essentially we use ResizeToContents until the widget is shown to the
user. That is not optimal, since this still can lead to a absurd size
for the keys column, but since the user can then resize it, it is far
better than before. Note: Each hide() + show() will resize the columns
again to their contents.

Task-number: QTCREATORBUG-8982
Change-Id: Iaccd32d4a2ce618a98f9b9d48141fd753c6f8b97
Reviewed-by: Tobias Hunger <tobias.hunger@digia.com>