QNX: Fixed paths to bar package to match build directory
Unless the package path, or application descriptor path, are changed
actively by the user, those values will now be automatically generated
from the build directory and source directory respectively.

This fixes a problem where the path to the package would be in
the debug build directory, even though the build mode was switched
to release.

Task-number: QTCREATORBUG-8991
Change-Id: I78137fc8a6ba0532663808ac9800c03e8371e148
Reviewed-by: Mehdi Fekari <mfekari@rim.com>
Reviewed-by: Nicolas Arnaud-Cormos <nicolas@kdab.com>
Reviewed-by: Tobias Hunger <tobias.hunger@digia.com>