ClearCase: Fix: Status actions was not correct for unix
A set view on unix will get / as topLevel. The output of
"cleartool -ls" with / as topLevel listed files with absolute path,
while with relative path for the the other cases (working directory
view and on windows). Root as topLevel is not 100% correct and should
be fixed in another patch. However, this patch fixes an annoying bug
for unix. (Before this patch multiple entries for the same file was
inserted in the status map - one with relative path and one with
absolute. Resulting in wrong lookup in updateStatusActions and thus
wrong status on the active file.

Change-Id: I178bafd21d712a445aca8e60ae3346549b55faaf
Reviewed-by: Orgad Shaneh <orgads@gmail.com>