Fix CMakeProjectManager Ninja CXXFLAGS detection
In recent CMake versions, a new backend, Ninja, was added with the intention
of providing an alternative to makefiles.  However, this backend stores the
CXXFLAGS for the project in a different format than it does for the makefiles
backends.  As the CMakeProjectManager assumes a makefile backend, it fails
to correctly obtain the CXXFLAGS of the project when Ninja is used.
This small patch will attempt to read CXXFLAGS from the build.ninja file
to correct this issue.

Task-number: QTCREATORBUG-9047

Change-Id: Ic19ae3b4802c7a12b5d0a92a7f1a2254bec5a3d2
Reviewed-by: Daniel Teske <daniel.teske@digia.com>