Git: bug fix for diff and log in BranchesDialog
Branch log and diff were sometimes having problems.
I have project in main git module, and I was not able to get branch diff and log in a submodule,
because they open new editor and as soon as it is opened m_repository changes to the master module.
Since repository working dir is set after editor opens, it is already changed in branchModel
Branch diff and log should now work as expected.

Change-Id: I72f730d77b8fe23ef6eac757415fe96321ef147b
Reviewed-by: Tobias Hunger <tobias.hunger@digia.com>