Revert "Add Core as dependency for the qtcreator executable in the qbs build."
This reverts commit a9754611f667d252ac015f2c8260caaf8dfbc909.

That patch does not work as-is, because it introduces a dependency from
the qtcreator library to the Core plugin without adding the respective
directory to the list of rpaths. While doing just that would make Creator
start up again, it seems conceptually wrong, because the Core plugin is
explicitly loaded by the application itself. Removing the dependency
altogether seems like the cleaner solution overall.

Change-Id: Ife7b726449701a0a61eb48d98c04c55a55cf47e3
Reviewed-by: Orgad Shaneh <orgads@gmail.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>