QmlDesigner.MetaInfo: avoid endless recusion into "." properties
Actually "." properties can recurse infinitely.
If a type MyType has the property parenType of MyType and
it is a read only pointer, we get an infinitely recusion into
"." properties.

parentType.parentType.parentType...

Actually this seems to be valid QML and is really the case with Menu.
(In the private API though, indicated by "__")

Since we enumerate "." properties in the Qt Quick Designer, we
run into an infinite loop.

I just cut the recursion at level 3.
So we stop at myParent.myParent.myParent which I think is reasonable.

Change-Id: I80866ace00f940000407cc25cec9ad6fac8b6fee
Reviewed-by: Fawzi Mohamed <fawzi.mohamed@digia.com>