ProjectExplorer/QNX: Add build step for checking device existence
This introduces a DeviceCheckBuildStep which any DeployConfigurationFactory
could use to insert a step to check whether a device is set up for the
current kit or not. Currently only used in the QNX deploy configuration.

If there is no device known to a QNX kit, ask the user
to add one. Launch the appropriate device wizard if the user agrees
to add a device.

Change-Id: I42db995f50890e2a2dd7aacc24a0049cdd6a6fee
Reviewed-by: Christian Kandeler <christian.kandeler@digia.com>