QmlDesigner.MetaInfo: C++ class names might equal QML type names
If the C++ class name equals the QML type name we looked up the
CppComponentValue of the C++ class.
This one has no version information, which creates issues
in the item library.

As a solution we try to resolve the type as a C++ class as a last resort.
Any QML type has precedence.

Task-number: QTCREATORBUG-9422
Change-Id: Ia84663cab1b4011f1ef6494de86ea6d92ab74bef
Reviewed-by: Fawzi Mohamed <fawzi.mohamed@digia.com>