FileSaver: Do not rollback if we never opened the file
This will silence some warnings about removing files with empty
filenames which we get due to trying to remove a file we never
managed to open.

Calling finalize() unconditionally is the right thing to do here
as that processes and/or displays on error message to the user
if there is any.

Change-Id: Ibe7d07233f2c2698e53a7ad21041bc5cc4fcac0e
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>