Git - Get changes actions dialog back to where it was
This dialog was before, in QtCreator 2.7, used only for show. It was in
main git menu, and available always. Although a lot of new options were
added to the dialog, I think it is good to get it back to where it was
before, so that 2.7 users can find it easier.

Also, I prefer this location for it, since it is easier to access here.

Change-Id: I510ab0efb081194e70f6297125012034103e13e1
Reviewed-by: Orgad Shaneh <orgads@gmail.com>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@digia.com>
Reviewed-by: Tobias Hunger <tobias.hunger@digia.com>