QmlDesigner.Rewriter: fixing encoding issue for Qt 4
QString uses utf-16 internally. If we explicitly convert to utf-8
and then use the QByteArray implicit constructor of QString again,
Qt 4 will interpret the utf-8 encoded string as latin1.
This results in a wrong string for everything that is not pure ASCII.

With Qt 5 it works, since the assumed default encoding of QByteArray
is utf-8. Still any conversion does not make sense.

Task-number: QTCREATORBUG-9493
Change-Id: Ifbaa4684e85d0ccd677956eeff88b4cc7d807922
Reviewed-by: Fawzi Mohamed <fawzi.mohamed@digia.com>