Qt4Project::updateRunConfigurations: Do not preserve unconfigured rcs
In 0ce0a053 code was added to ensure that if there is a configured
runconfiguration, then it'll be made active. This fixes that code to
work more robustly.

Consider the case:
Two runconfigurations
a) custom executable runconfiguration unconfigured (active)
b) configured qt4runconfiguration

The code in upadteRunConfigurations would put a) into the
existingConfigured list since it is active, b) since it fits.

Thus the code at the end would not change the active runconfiguration,
since existingConfigured.at(0)->isConfigured() would be false.

Consider now the same case, but with the reverse order.
Now the code makes b) active. The ordering between rcs should not
matter.

Also consider that in the second case, the next run of
updateRunConfigurations will remove the now non active unconfigured
runconfiguration.

Change-Id: I98291db4eda22e9dae3e493a8f31c7d9bb5420d6
Reviewed-by: Tobias Hunger <tobias.hunger@digia.com>