Qt4Project: On folder change just schedule a async parse
On a folder change we need to collect all files which we add to the project
tree due to INSTALL or DEPLOYMENT rules. On a folder change the old
implementation only collected the files in the changed folder, but also
started a updateCodeModels(). If multiple Qt4PriFileNodes would react to
the folderChanged() signal, then a lot of updateCodeModels() would be
started.

This new code does a unnecessary complete iteration of the files starting
from the top level INSTALLS or DEPLOYMENT folders, but optimizes the
updateCodeModels() by going through the more advanced scheduling the
qt4project does for evaluation.

Task-number: QTCREATORBUG-9581
Change-Id: I50345516c56bd33e6e1e1b2e0d7db50635ea1552
Reviewed-by: Tobias Hunger <tobias.hunger@digia.com>
Reviewed-by: Nikolai Kosjar <nikolai.kosjar@digia.com>