Workaround for Clang generating wrong debug info
Clang obviously takes PWD as a basis for source paths in debug
information (AT_comp_dir) if it is set. Since the build tools are not
executed in a shell that would set PWD to a sensible value, debug
information with Clang is broken if the build environment got a wrong
PWD to start with, for example by starting Qt Creator from a shell
(which sets PWD).

Workaround is to explicitly unset PWD for the build environment used for
the Clang tool chain.

Task-number: QTCREATORBUG-8825

Change-Id: Ia3e47693121edd61b0ea43bddbb1fcaaac9eb390
Reviewed-by: Daniel Teske <daniel.teske@digia.com>