Don't accept a column/line number selection just because of ":"
If you had typed "filename.cpp:142" in the Locator field, Creator
would offer "Column 142" the first option. At the very least, that's
confusing because the user intended to select line 142, not column --
for a while I thought someone had typed the wrong word in the source
code.

But even worse, this selection was the first entry, so instead of
jumping to filename.cpp on line 142, the user ended up still on the
same file and line, but on a different column.

Instead, simply validate that the parts before and after the colon are
actually numbers.

Change-Id: I1f1fbd33d702ae26c41fd2485aa99fae800e1410
Reviewed-by: Eike Ziller <eike.ziller@digia.com>