Generic Android error messages
We want to give better feedback for installation problems and
other issues. There may still be undiscovered ways to mess up the
creator configuration, and we might find them after the string
freeze.

This adds a number of strings that we can use later.
We can add new error codes after the string freeze (and even in
patch releases). This means that even if we can't give people
detailed help for their problem in the UI, at least we give them
a googleable string.

We can then add better messages for the new errors in the next minor
release.

Change-Id: Ida7921fce5142d55144b0ba022de04726872187a
Reviewed-by: Daniel Teske <daniel.teske@digia.com>