QmlDesigner.PropertyEditor: extend lookup rules for extensions
If specifics for pane .qml files are located inside the plugin
in the designer sudirectory, we still required the path to be fully
qualified.
e.g. myComponnents/designer/myComponent/MySpecifics.qml

This is unconvinient and also a behaivour change for pure directory imports.
Pure directory imports were not fully qualified until recently.

With this patch we also allow:
myComponnents/designer/MySpecifics.qml

Change-Id: I9a61016f425caf5d7d15140554c72f4a69a6687a
Reviewed-by: Marco Bubke <marco.bubke@digia.com>
Reviewed-by: Eike Ziller <eike.ziller@digia.com>