Progress summary: Forward click to "current" progress details.
It was annoying that it wasn't possible to click on the summary progress
bar to e.g. open the compile output or issues pane, even when you had
only a build running.

This change forwards the click to either the progress that currently
shows a details widget in the status bar (like the build errors/warnings
from the build progress), or to the last progress that was started.

Change-Id: I4b50dd4b43bdc12fac329d33e158fa759bee4d28
Reviewed-by: Tobias Hunger <tobias.hunger@digia.com>