Fix wrong progress when loading session.
The code was trying to show one tick for each editor that is opened, but
editors are actually only opened for a file, if it is visible in a view.
So editorOpened is only sent for one editor per open split.
The new code just shows one tick for restoring editors, to keep things
simple.

Change-Id: I6a24061858f532c3ccfded9677b8928ae62a1448
Reviewed-by: Daniel Teske <daniel.teske@digia.com>