Wizards.Dialog: fix the All Templates case
The "All Templates" case in the comobox did not work as expected
from users.
It still did only show wizards we have a kit installed for
and it was not available for the case of one platform.

The semantics for "All Templates" were actually "All available Platforms".
This was confusing especially in the case of no platform.
Now "All Templates" does show all templates, but the wizard give an error
message, if no corresponding kit is available.

Task-number: QTCREATORBUG-9792
Change-Id: Ia83a0577bde7e726b2638868bcaffbe13499f2f7
Reviewed-by: Eike Ziller <eike.ziller@digia.com>