Qnx: Fixing UI freeze when canceling SSH-keys generation
The UI freeze happens when an user cancels 'Add BlackBerry Device' wizard
while SSH-keys are being generated.

Removing a need for terminate() and wait() calls in
BlackBerryDeviceConfigurationWizardSshKeyPage dtor by invoking a new
thread each time a Generate button is pressed. Such a thread can delete
itself later when the calculation is finished.
The only drawback is that the calculation is still happening even when the
wizard is closed. Just the calculation results are not used anywhere.

Task-number: QTCREATORBUG-9888

Change-Id: Ibc5e98d08e129d7f76620a5dea9ed5190932243b
Reviewed-by: Mehdi Fekari <mfekari@blackberry.com>
Reviewed-by: Tobias Hunger <tobias.hunger@digia.com>