Debugging: Continue Cpp debugging even if QML debugger fails
In mixed debugging, QML debugger is the slave engine. Continue
debugging with the master engine i.e. Cpp engine even if the
slave engine fails. The slave engine may fail when the application
is not set up for QML/JS debugging or when there are insufficient
available ports etc.

Task-number: QTCREATORBUG-9836

Change-Id: I2549196413c839e660925fe2edb3182e6848b5fc
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>
Reviewed-by: Christiaan Janssen <christiaan.janssen@digia.com>
Reviewed-by: Kai Koehne <kai.koehne@digia.com>