Qnx: update the bbnk-env parser for internal NDKs
In recent internal NDK versions (10.2) some new
environment variables are introduced (they are evaluated
when soucring the bbnk-script) that QNX_TARGET and QNX_HOST
are depending on.

This patch is a workaroud, in the future the parser
should be deprecated/replaced to make the plugin less depending on
the bbndk-scrit changes.

Change-Id: I8d2dc7f0a87cbada6f8807df2bafdfd14476e01d
Reviewed-by: Tobias Nätterlund <tobias.naetterlund@kdab.com>
Reviewed-by: Nicolas Arnaud-Cormos <nicolas@kdab.com>