Android: Clean up before packaging, be a little bit more careful.
The android/libs folder contains both the library from the project and
libraries from qt. This adds a small check before removing libQt5*.so,
that checks if that file is in qt's lib folder. That fixes the bug
below, but e.g. naming your project Qt5Test won't work.
Long term we'll need to overhaul the packaging and make that more
robust.

Task-number: QTCREATORBUG-9895
Change-Id: I48f735114ee95668a862549cfe1537aeab968dfb
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@digia.com>