Qnx: Auto-generating SSH key used for connecting to a BB10 device
BlackBerryDeviceConfigurationWizardSshKeyPage wizard page has been removed.
It is not needed anymore. The wizard automatically generates or reuses
a SSH key that is stores in __settings__/qnx/id_rsa file. The same SSH key
is used for connecting to all BB10 devices.
The generation of 'default' SSH key is gone as a second step of
BlackBerryDeviceConfigurationWizardQueryPage wizard page.

Disabling device-tester for BlackBerryDeviceConfiguration.
Task-number: QTCREATORBUG-9977

Change-Id: Ice6068530c9c72fa82907decaaa3dca6077e9c3a
Reviewed-by: Tobias Hunger <tobias.hunger@digia.com>
Reviewed-by: David Kaspar <dkaspar@blackberry.com>