CppEditor: fix showing 'Type Hierarchy'
When we open project and do not open any file, then open 'Type Hierarchy'
(from combobox) we see "No type hierarchy available"(and this is ok). Then
if we open any file and we try to see 'Type Hierarchy'(Ctrl+Shift+T) then we
still see "No type hierarchy available"(and this is not ok).
It will also happens if we save a session with opened "Type Hierarchy".

This change fixes this. Additionally when we select something for which there
is no 'Type Hierarchy' then it displays "No type hierarchy available".

Task-number: QTCREATORBUG-9819
Change-Id: Ib6a152f481057098f3d8a4335bb2d4a31fc5e1ef
Reviewed-by: Tobias Hunger <tobias.hunger@digia.com>
Reviewed-by: Nikolai Kosjar <nikolai.kosjar@digia.com>